requirements.txt
ftx
ciso8601
python-dotenv
